+++ packages/synapse-interface/pages/state-managed-bridge/index.tsx
@@ -11,7 +11,6 @@ import {
   getPublicClient,
   waitForTransactionReceipt,
 } from '@wagmi/core'
-
 import { InputContainer } from '@/components/StateManagedBridge/InputContainer'
 import { OutputContainer } from '@/components/StateManagedBridge/OutputContainer'
 import { BridgeExchangeRateInfo } from '@/components/StateManagedBridge/BridgeExchangeRateInfo'
@@ -54,7 +53,6 @@ import { Token } from '@/utils/types'
 import { txErrorHandler } from '@/utils/txErrorHandler'
 import { approveToken } from '@/utils/approveToken'
 import { stringToBigInt } from '@/utils/bigint/format'
-
 import { fetchAndStoreSingleNetworkPortfolioBalances } from '@/slices/portfolio/hooks'
 import {
   updatePendingBridgeTransaction,
@@ -65,6 +63,7 @@ import { useAppDispatch } from '@/store/hooks'
 import { RootState } from '@/store/store'
 import { getTimeMinutesFromNow } from '@/utils/time'
 import { isTransactionReceiptError } from '@/utils/isTransactionReceiptError'
+import { isTransactionUserRejectedError } from '@/utils/isTransactionUserRejectedError'
 import {
   MaintenanceWarningMessages,
   useMaintenanceCountdownProgresses,
@@ -74,6 +73,7 @@ import {
   getBridgeModuleNames,
 } from '@/components/Maintenance/Maintenance'
 import { wagmiConfig } from '@/wagmiConfig'
+import { useStaleQuoteRefresher } from '@/utils/hooks/useStaleQuoteRefresher'
 
 const StateManagedBridge = () => {
   const { address } = useAccount()
@@ -93,12 +93,14 @@ const StateManagedBridge = () => {
     bridgeQuote,
     debouncedFromValue,
     destinationAddress,
+    isLoading: isQuoteLoading,
   }: BridgeState = useBridgeState()
   const { showSettingsSlideOver, showDestinationAddress } = useSelector(
     (state: RootState) => state.bridgeDisplay
   )
 
-  const [isApproved, setIsApproved] = useState(false)
+  const [isWalletPending, setIsWalletPending] = useState<boolean>(false)
+  const [isApproved, setIsApproved] = useState<boolean>(false)
 
   const dispatch = useAppDispatch()
 
@@ -153,6 +155,7 @@ const StateManagedBridge = () => {
 
     try {
       dispatch(setIsLoading(true))
+      const currentTimestamp: number = getTimeMinutesFromNow(0)
 
       const allQuotes = await synapseSDK.allBridgeQuotes(
         fromChainId,
@@ -273,6 +276,7 @@ const StateManagedBridge = () => {
             estimatedTime: estimatedTime,
             bridgeModuleName: bridgeModuleName,
             gasDropAmount: BigInt(gasDropAmount.toString()),
+            timestamp: currentTimestamp,
           })
         )
 
@@ -313,6 +317,13 @@ const StateManagedBridge = () => {
       }
     }
   }
+
+  useStaleQuoteRefresher(
+    bridgeQuote,
+    getAndSetBridgeQuote,
+    isQuoteLoading,
+    isWalletPending
+  )
 
   const approveTxn = async () => {
     try {
@@ -366,6 +377,7 @@ const StateManagedBridge = () => {
       })
     )
     try {
+      setIsWalletPending(true)
       const wallet = await getWalletClient(wagmiConfig, {
         chainId: fromChainId,
       })
@@ -496,6 +508,10 @@ const StateManagedBridge = () => {
       console.log('Error executing bridge', error)
       toast.dismiss(pendingPopup)
 
+      if (isTransactionUserRejectedError(error)) {
+        getAndSetBridgeQuote()
+      }
+
       /** Fetch balances if await transaction receipt times out */
       if (isTransactionReceiptError(error)) {
         dispatch(
@@ -507,6 +523,8 @@ const StateManagedBridge = () => {
       }
 
       return txErrorHandler(error)
+    } finally {
+      setIsWalletPending(false)
     }
   }